added code runner to PyProm_common.
