+++ PyProma_GUI/PyProma_dirview/pyproma_dirview.py
@@ -279,38 +279,6 @@ class DirView:
             path = os.path.normpath(path)
             pyperclip.copy(path)
 
-    @staticmethod
-    def code_runner(command: str | list):
-        """this func runs bash command and shows outputs to textbox.
-
-        Args:
-            command (str): command
-        """
-        root = tk.Toplevel()
-        root.title("code runner")
-        text = tk.Text(root)
-        text.pack()
-        try:
-            process = subprocess.Popen(
-                command, shell=False, text=True,
-                stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-
-            while True:
-                output = process.stdout.readline()
-                if output == "":
-                    break
-                text.insert(tk.END, output)
-                text.see(tk.END)
-        except subprocess.CalledProcessError as e:
-            messagebox.showerror(
-                parent=root,
-                title="subprocess.CalledProcessError",
-                message=str(e))
-        else:
-            messagebox.showinfo(parent=root, message="Command succeed.")
-
-        root.destroy()
-
     def dir_menu_on_right_click(self, event: tk.Event):
         """this func shows right-clicked menu.
